rf - extracted DirectoryLister fo reuse and ext
